adding some changes
