+++ lib/main.dart
@@ -6,7 +6,7 @@ import 'package:get/get_navigation/src/root/get_material_app.dart';
 import 'package:taskmanagement/screen/homescreen/home_view.dart';
 import 'package:taskmanagement/screen/profile/edit_profile_view.dart';
 import 'package:taskmanagement/screen/profile/show_profile.dart';
-import 'package:taskmanagement/screen/signup_screen.dart';
+import 'package:taskmanagement/screen/signupscreen/signup_screen.dart';
 import 'package:taskmanagement/screen/splash/splash_view.dart';
 import 'package:taskmanagement/services/authentication.dart';
 import 'package:taskmanagement/utils/theme/theme.dart';
