+++ lib/screen/taskscreen/description.dart
@@ -1,14 +1,33 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:taskmanagement/constant/colors.dart';
 
+// class forgotpasswordbuttonwidget extends StatelessWidget {
+//   const forgotpasswordbuttonwidget({
+//     super.key,
+//     required this.btnIcon,
+//     required this.title,
+//     required this.subTitle,
+//     required this.onTap,
+//   });
+//   final IconData btnIcon;
+//   final String title,subTitle;
+//   final VoidCallback onTap;
 class Description extends StatelessWidget {
-  final String? title, description;
-
-  const Description({required Key key, this.title, this.description}) : super(key: key);
+  const Description({
+    super.key,
+    this.title,
+    this.description,
+  });
+  final String? title;
+  final String? description;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Description')),
+      appBar: AppBar(
+        title: Text('Description'),
+        backgroundColor: tSecondaryColor,
+      ),
       body: Container(
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
