+++ lib/services/authentication.dart
@@ -2,11 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:get/get.dart';
 import 'package:taskmanagement/screen/homescreen/home_view.dart';
-import 'package:taskmanagement/screen/signup_screen.dart';
-
-import '../screen/addtask.dart';
-import '../screen/taskscreen.dart';
-
+import 'package:taskmanagement/screen/signupscreen/signup_screen.dart';
 
 class AuthenticationRepository extends GetxController {
   static AuthenticationRepository get instance => Get.find();
@@ -30,24 +26,29 @@ class AuthenticationRepository extends GetxController {
   //creating instance of firebaseauth
 
   //getting the current login user
-  User? currentUser= FirebaseAuth.instance.currentUser;
+  User? currentUser = FirebaseAuth.instance.currentUser;
   //defining method to authenticate the user
-  Future<void> createUserWithEmailAndPassword(String fullName,String email,
-      String password,String phoneNo) async {
+  Future<void> createUserWithEmailAndPassword(
+      String fullName, String email, String password, String phoneNo) async {
     try {
       //query for the authentication
-      await _auth.createUserWithEmailAndPassword(email: email, password: password);
+      await _auth.createUserWithEmailAndPassword(
+          email: email, password: password);
       // print("user created");
-      Get.snackbar('successful','you logged in successfully' );
+      Get.snackbar('successful', 'you logged in successfully');
       //query to store user data in firestore
-      Get.to(()=>Home_view());
-      FirebaseFirestore.instance.collection('users').doc().set({
-        "FullName":fullName,
-        "Email":email,
-        "Password":password,
-        "PhoneNo":phoneNo,
-        "UserId":currentUser!.uid,
-      });
+      Get.to(() => Home_view());
+      try {
+        FirebaseFirestore.instance.collection('users').add({
+          "FullName": fullName,
+          "Email": email,
+          "Password": password,
+          "PhoneNo": phoneNo,
+          "UserId": currentUser!.uid,
+        });
+      } catch (exception) {
+        Get.snackbar("error", "error in saving data");
+      }
     } on FirebaseAuthException catch (e) {
       if (e.code == 'weak-password') {
         Get.snackbar("Error", "weak password",
@@ -57,9 +58,5 @@ class AuthenticationRepository extends GetxController {
             snackPosition: SnackPosition.BOTTOM);
       }
     }
-
   }
-
-
-
 }
